# [v1.x] provide a faster PrefetchedDataLoader